backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "fastapi backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncmy>=0.2.10",
    "clerk-backend-api>=3.0.5",
    "fastapi>=0.116.1",
    "openai>=1.95.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "svix>=1.69.0",
    "uvicorn>=0.35.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
